# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Portico
  version: 0.1.0
  summary: The management and self-service API.
  description: |
    Portico's own HTTP API, under `/api/v1`.

    **What is deliberately not here.** SCIM 2.0 is defined by RFC 7643 and
    RFC 7644, and OpenID Connect, OAuth 2.1, SAML 2.0, and CAS are defined by
    their own specifications. Restating them in house style would produce a
    second description that can disagree with the first, and clients for
    those protocols read the specification rather than this file. `/api/v1`
    is the surface that has no external definition, so it is the surface this
    document covers. See `docs/scim.md` and `docs/federation.md` for the
    rest, and `docs/webhooks.md` for the outbound side, which is a request
    Portico makes rather than one it serves.

    **Every response is wrapped**, including failures — see the `Envelope`
    schema. A client reads `code` to decide what happened; the HTTP status
    agrees with it but is coarser. `data` is always present and is `null`
    when there is no result.

    **Every request acts inside exactly one tenant.** For an authenticated
    call the tenant comes from the token and the `X-Tenant` header is
    ignored; sending one that disagrees changes nothing. Only sign-in,
    registration, and the other endpoints reachable without a token take the
    tenant from the header or the body, because there is no token yet.

    This file is checked against the running router by
...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/tenants:
    get:
      tags: [Tenants]
      operationId: listTenants
      summary: Every tenant on this deployment, with a count of what is in it.
      description: |
        The one endpoint in this API that sees past the caller's own tenant,
        and the narrowest crossing that answers the question: sizes, never
        contents.

        It exists only where the deployment set `PORTICO_TENANT_CONSOLE`, and
        then only for an administrator of the `default` tenant. Everybody else
        — including an administrator of another tenant — is answered `404`,
        which is deliberately the same answer a deployment without the feature
        gives: a `403` would confirm that the console exists and that somebody
        else has it.
      responses:
        "200":
          description: The tenants.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TenantOverviewListEnvelope" }
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/tenants/{code}/status:
    put:
      tags: [Tenants]
      operationId: setTenantStatus
      summary: Disable a tenant, or turn one back on.
      description: |
        Disabling refuses every sign-in to that tenant and ends the sessions it
        already had. Nothing is deleted and enabling restores it.

        `confirm` must equal the tenant's own code. The check is here rather
        than only in a dialog, because a caller that draws no dialog would
        otherwise be one request away from signing out everybody in a tenant.

        The `default` tenant cannot be disabled through this endpoint: it is
        what serves the console, and there would be no way back from a browser.
      parameters:
        - name: code
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status, confirm]
              properties:
                status:
                  type: string
                  enum: [ACTIVE, DISABLED]
                confirm:
                  type: string
                  description: The tenant's own code, typed back.
      responses:
        "200":
          description: The tenant, as it now is.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TenantEnvelope" }
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        The token from `/api/v1/auth/login`. It carries the tenant, which is
        why authenticated requests do not take one.

        SCIM credentials are also bearer tokens but are not interchangeable
        with these: one authenticates a directory, the other a person, and
        each is refused where the other belongs.
...
  responses:
...
    NotFound:
      description: |
        No such record in this tenant. A record that exists in another tenant
        is reported the same way, because saying otherwise would confirm it
        exists.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Envelope" }
...
  schemas:
...
    Envelope:
      type: object
      required: [code, message, data]
      description: |
        Every response, successful or not, has this shape. `code` is
        `SUCCESS` or a stable SCREAMING_SNAKE_CASE identifier; clients branch
        on it rather than on the message, which is prose meant for a person
        and is translated.
      properties:
        code:
          type: string
          examples: [SUCCESS, USER_NOT_FOUND]
        message:
          type: string
          description: Human-readable, and safe to show to an end user.
        data:
          description: Null when there is no result — never `{}` or `[]`.
...
    TenantOverview:
      type: object
      description: |
        A tenant and how much is inside it. Counts and nothing else: an
        operator may learn a tenant's size and never its contents, so there is
        no name, address, or identifier here belonging to anybody in it.
      properties:
        id: { type: string }
        code: { type: string }
        name: { type: string }
        status: { type: string, enum: [ACTIVE, DISABLED] }
        createdAt: { type: string, format: date-time }
        users: { type: integer }
        activeUsers: { type: integer }
        organizations: { type: integer }
        applications:
          type: integer
          description: OAuth clients, SAML service providers and CAS services together.
        lastActivity:
          type: string
          format: date-time
          nullable: true
          description: |
            When anything was last recorded in that tenant's audit trail. Null
            for a tenant nothing has ever happened in, which is usually the row
            worth looking at.
...
    TenantOverviewListEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              type: array
              items: { $ref: "#/components/schemas/TenantOverview" }
...